+++ openfl-workspace/keras/jax/mnist/plan/defaults
