+++ openfl-workspace/keras/jax/mnist/requirements.txt
