+++ openfl-workspace/keras/jax/mnist/src/mnist_utils.py
